requirements.txt
flask
werkzeug
ultralytics
opencv-python-headless
numpy
Pillow
SQLAlchemy
Flask-SQLAlchemy
